utils/configuration_validator/requirements.txt
pytest>=2.6.4
PyYAML>=3.11
fuel-tasks-validator==0.01
